Fix compiler errors and convert healthData to struct instead of tuple

+++ contracts/storage/Margin.sol
@@ -262,7 +262,7 @@ library Margin {
                     collateralUsd.toInt() +
                         position.getPnl(marketPrice) +
                         position.getAccruedFunding(market, marketPrice) -
-                        position.getAccruedUtilization(market, marketPrice) -
+                        position.getAccruedUtilization(market, marketPrice).toInt() -
                         position.accruedFeesUsd.toInt(),
                     0
                 )
